# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Palace Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  isActive: boolean;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface ProductVariant {
  id: number;
  productId: number;
  name: string;
  price: number;
  isAvailable: boolean;
}

export interface ProductVariantInput {
  name: string;
  price: number;
  isAvailable?: boolean;
}

export interface ProductAddon {
  id: number;
  productId: number;
  name: string;
  price: number;
  isAvailable: boolean;
}

export interface ProductAddonInput {
  name: string;
  price: number;
  isAvailable?: boolean;
}

export interface Product {
  id: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  basePrice: number;
  isAvailable: boolean;
  isFeatured: boolean;
  sortOrder?: number;
  variants?: ProductVariant[];
  addons?: ProductAddon[];
  createdAt: string;
}

export interface ProductInput {
  categoryId: number;
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  basePrice: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
}

export interface ProductUpdate {
  categoryId?: number;
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
  basePrice?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
}

export interface DeliveryLocation {
  id: number;
  zoneId: number;
  name: string;
}

export interface DeliveryZone {
  id: number;
  name: string;
  fee: number;
  isActive: boolean;
  locations?: DeliveryLocation[];
  createdAt: string;
}

export interface DeliveryZoneInput {
  name: string;
  fee: number;
  isActive?: boolean;
  locations?: string[];
}

export interface DeliveryZoneUpdate {
  name?: string;
  fee?: number;
  isActive?: boolean;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  /** @nullable */
  variantName?: string | null;
  /** @nullable */
  addonNames?: string | null;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
}

export interface OrderItemInput {
  productId: number;
  productName: string;
  variantId?: number;
  variantName?: string;
  addonIds?: number[];
  addonNames?: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  notes?: string;
}

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  customerName?: string;
  customerPhone?: string;
  /** @nullable */
  customerEmail?: string | null;
  status: string;
  paymentMethod: string;
  paymentStatus: string;
  subtotal: number;
  deliveryFee: number;
  total: number;
  deliveryAddress: string;
  /** @nullable */
  deliveryLocation?: string | null;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  deliveryNotes?: string | null;
  /** @nullable */
  estimatedDeliveryTime?: string | null;
  items?: OrderItem[];
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  paymentMethod: string;
  subtotal: number;
  deliveryFee: number;
  total: number;
  deliveryAddress: string;
  deliveryLocation?: string;
  zoneId?: number;
  deliveryNotes?: string;
  items: OrderItemInput[];
}

export interface OrderStatusUpdate {
  status: string;
  estimatedDeliveryTime?: string;
}

export interface PaymentConfirmInput {
  action: string;
  paymentStatus?: string;
}

export interface Favorite {
  id: number;
  userId: number;
  productId: number;
  product?: Product;
  createdAt: string;
}

export interface RestaurantSettings {
  id: number;
  restaurantName: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  isOpen: boolean;
  /** @nullable */
  openingHours?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccountName?: string | null;
  /** @nullable */
  bankAccountNumber?: string | null;
  /** @nullable */
  estimatedDeliveryTime?: string | null;
  createdAt: string;
}

export interface RestaurantSettingsUpdate {
  restaurantName?: string;
  tagline?: string;
  phone?: string;
  email?: string;
  address?: string;
  logoUrl?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  isOpen?: boolean;
  openingHours?: string;
  whatsappNumber?: string;
  instagramUrl?: string;
  facebookUrl?: string;
  twitterUrl?: string;
  bankName?: string;
  bankAccountName?: string;
  bankAccountNumber?: string;
  estimatedDeliveryTime?: string;
}

export interface TopProduct {
  productName: string;
  orderCount: number;
  revenue: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface Analytics {
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  todayOrders: number;
  todayRevenue: number;
  topProducts: TopProduct[];
  ordersByStatus?: StatusCount[];
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
available?: boolean;
};

export type ListOrdersParams = {
status?: string;
userId?: number;
};