# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Palace Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductAddon } from './productAddon';
import type { ProductVariant } from './productVariant';

export interface Product {
  id: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  basePrice: number;
  isAvailable: boolean;
  isFeatured: boolean;
  sortOrder?: number;
  variants?: ProductVariant[];
  addons?: ProductAddon[];
  createdAt: Date;
}